+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Hovering a course block in the schedule grid now highlights the corresponding course in the sidebar
+- Schedule sidebar now groups courses by term (A, B, C, D) with labeled separators
 
 ## [2026-03-31]